use the proper typedef for ssize_t (windows)

+++ miner.h
@@ -19,7 +19,8 @@ extern "C" {
 #define strdup(x) _strdup(x)
 #define strncasecmp(x,y,z) _strnicmp(x,y,z)
 #define strcasecmp(x,y) _stricmp(x,y)
-typedef int ssize_t;
+#include <BaseTsd.h>
+typedef SSIZE_T ssize_t;
 #undef HAVE_ALLOCA_H
 #undef HAVE_SYSLOG_H
 #endif
